refactor(app): extract session window controller from layout.tsx
Move the session window data layer (window store, load + reconcile, derived
session/section/navigation memos, SDK session.created/updated/deleted sync,
show-more) out of pages/layout.tsx into
pages/layout/pawwork-session-controller.ts as createPawworkSessionController.

Mechanical extraction: reactive deps passed as a named dependency object;
memos / effect / event listener run under the Layout owner unchanged. layout.tsx
consumes the returned accessors. Updated the timestamp-cache source-assertion
test to point at the controller file (code moved verbatim, behavior identical).

layout.tsx 1960 -> 1704 lines.

Refs #1056
